not working

+++ tutorial/query.sql.go
@@ -34,7 +34,7 @@ DELETE FROM authors
 WHERE id = $1
 `
 
-func (q *Queries) DeleteAuthor(ctx context.Context, id int64) error {
+func (q *Queries) DeleteAuthor(ctx context.Context, id int32) error {
 	_, err := q.db.ExecContext(ctx, deleteAuthor, id)
 	return err
 }
@@ -44,7 +44,8 @@ SELECT id, name, bio FROM authors
 WHERE id = $1 LIMIT 1
 `
 
-func (q *Queries) GetAuthor(ctx context.Context, id int64) (Author, error) {
+//WHY IS IT NOT CREATING authors as a table??
+func (q *Queries) GetAuthor(ctx context.Context, id int32) (Author, error) {
 	row := q.db.QueryRowContext(ctx, getAuthor, id)
 	var i Author
 	err := row.Scan(&i.ID, &i.Name, &i.Bio)
